[LOG ENTRY] Moved Orginal Controller To More Specific Namespace
