+++ app/views/projects/subprojects/log_entries/edit.turbo_stream.erb
