+++ app/views/projects/subprojects/log_entries/new.turbo_stream.erb
