+++ app/views/projects/subprojects/log_entries/show.turbo_stream.erb
