+++ app/views/users/show.html.erb
@@ -37,7 +37,7 @@
           <div class="flex flex-row gap-2 items-center">
             <%= render ActionButtonComponent.new(to: project_subproject_log_entry_path(log_entry.subproject.project, log_entry.subproject, log_entry), icon: "view", turbo_stream: true) %>
             <%= render ActionButtonComponent.new(to: edit_project_subproject_log_entry_path(log_entry.subproject.project, log_entry.subproject, log_entry), icon: "edit", turbo_stream: true, colour: :info) %>
-            <%= render ActionButtonComponent.new(to: project_subproject_log_entry_path(log_entry.subproject.project, log_entry.subproject, log_entry), method: :delete, icon: "delete", colour: :error, confirm: t("log_entries.destroy.confirm")) %>
+            <%= render ActionButtonComponent.new(to: project_subproject_log_entry_path(log_entry.subproject.project, log_entry.subproject, log_entry), method: :delete, icon: "delete", colour: :error, confirm: t("projects.subprojects.log_entries.destroy.confirm")) %>
           </div>
         <% end %>
       <% end %>
