--- config/locales/en/log_entries.en.yml
@@ -1,25 +1,0 @@
----
-en:
-  activerecord:
-    attributes:
-      log_entry:
-        created_at: Created At
-        project: Project
-        subproject: Subproject
-        updated_at: Updated At
-        user: User
-  log_entries:
-    create:
-      success: Log entry was successfully created.
-    destroy:
-      confirm: Are you sure you want to delete this log entry?
-      success: Log entry was successfully deleted.
-    edit:
-      title: Edit Log Entry
-    new:
-      title: New Log Entry
-    show:
-      close: Close
-      title: Log Entry Details
-    update:
-      success: Log entry was successfully updated.
